+++ Goap/Conditions/ConditionInterface.cs
@@ -3,6 +3,6 @@ namespace TsunagiModule.Goap
     public interface ConditionInterface
     {
         public bool IsSatisfied(State state);
-        public float EstimateDistance(State state);
+        public double EstimateDistance(State state);
     }
 }
